refactor: organize all linked list`s functions inside the lib
